requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
httpx>=0.27.0
requests>=2.31.0
pytest>=7.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.0.0
chromadb>=0.4.0
joblib>=1.3.0
statsmodels
openai
